requirements.txt
discord.py==2.2.2
python-dotenv==1.0.0
revChatGPT==4.2.0
openai==0.27.4
asgiref==3.6.0
GoogleBard==0.0.6.1
EdgeGPT==0.1.22.1
PyNaCl==1.4.0